+++ Release_Notes.md
@@ -33,6 +33,7 @@ Marked AbstractDatabaseFunctions and AbstractBaseModel as Obsolete
     - MSecurityEntity                       to AAddedUpdated
     - MState                                to AAddedUpdated
     - MLoggingProfile                       to ADatabaseTable
+    - UIFeedback                            to ADatabaseTable
 
 ### Known Bugs
 - There is an issue where the refresh token is being revoked when it shouldn't be causing a logout to occure